+++ CHANGES.md
@@ -1,7 +1,10 @@
 ## Unreleased
 
+## 7.2.0 (2024-10-04)
+
 * update BaseBackend to use a default coord_crs from the tms (author @AndrewAnnex, https://github.com/developmentseed/cogeo-mosaic/pull/234)
 * add python 3.12 support
+* Add tms parameter to cli for MosaicJSON (co-author @AndrewAnnex, https://github.com/developmentseed/cogeo-mosaic/pull/233)
 
 ## 7.1.0 (2023-12-06)